requirements.txt
langchain-groq
langchain-google-genai
langchain-openai
langchain-core
langchain-community
python-dotenv
pandas
pathlib
unstructured[md]
spacy
kagglehub
spotipy
git+https://github.com/johnwmillr/LyricsGenius.git